+++ src/VPinCommander.Mobile/AppShell.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace VPinCommander.Mobile;
+
+public partial class AppShell : Shell
+{
+	public AppShell()
+	{
+		InitializeComponent();
+	}
+}