+++ src/VPinCommander.Mobile/MainPage.xaml.cs
@@ -1,0 +1,10 @@
+namespace VPinCommander.Mobile;
+
+public partial class MainPage : ContentPage
+{
+    public MainPage()
+    {
+        InitializeComponent();
+        BindingContext = new CabinetsPageViewModel();
+    }
+}